Refactor for convert logic. Logic is now split into multiple files and folders. TS Export is now combined into one object.
