+++ README.md
@@ -32,7 +32,7 @@ Tokens are available as named SCSS variables and mixins with **\$fi-** prefix. T
 
 Color tokens are available as variables with additional **color-** prefix. Colors are provided as hsl values.
 
-Example from **tokens.scss**:
+Excerpt from **tokens.scss**:
 
 ```scss
 $fi-color-brand-base: hsl(214, 100%, 24%);
@@ -50,7 +50,7 @@ Example use case:
 
 Typography tokens are available as mixins with additional **text-** prefix.
 
-Example from **tokens.scss**:
+Excerpt from **tokens.scss**:
 
 ```scss
 @mixin fi-text-heading1 {
@@ -74,7 +74,7 @@ h1 {
 
 Spacing tokens are available as variables with additional **spacing-** prefix.
 
-Example from **tokens.scss**:
+Excerpt from **tokens.scss**:
 
 ```scss
 $fi-spacing-l: 32px;
@@ -90,64 +90,41 @@ Example use case:
 
 ## JavaScript and TypeScript
 
-Tokens are available as named **tokens** export and can be imported to JavaScript and TypeScript files.
+Tokens are available as named **suomifiDesignTokens** export and can be imported to JavaScript and TypeScript files.
+
+**suomifiDesignTokens** exports the design tokens as css formatted strings and in more granular format so that individual properties are available with values and units.
 
 JavaScript example:
 
 ```js
-import { tokens } from 'suomifi-design-tokens';
+import { suomifiDesignTokens } from 'suomifi-design-tokens';
 ```
 
 TypeScript example with typings:
 
 ```ts
-import { tokens, DesignTokens } from 'suomifi-design-tokens';
+import {
+  suomifiDesignTokens,
+  DesignTokens
+} from 'suomifi-design-tokens';
+
+const tokens: DesignTokens = suomifiDesignTokens;
 ```
 
 ### 🎨 Colors
 
-Color tokens are available with **colors** property. Colors are provided as both separated h, s and l values and as combined hsl string.
+Color tokens are available with **colors** property as css strings. Separated h, s and l values can be accessed through **values** property.
 
-Example from **tokens object**:
+Excerpt from **suomifiDesignTokens** object:
 
 ```js
-exports.tokens = {
+exports.suomifiDesignTokens = {
   colors: {
-    whiteBase: { hsl: 'hsl(0, 0%, 100%)', h: 0, s: 0, l: 100 }
-  }
-};
-```
-
-JavaScript example:
-
-```js
-const brandBaseHsl = tokens.colors.brandBase.hsl;
-```
-
-TypeScript example with typings:
-
-```ts
-import { ColorToken } from 'suomifi-design-tokens';
-
-const brandBase: ColorToken = tokens.colors.brandBase;
-const brandBaseHsl: string = brandBase.hsl;
-```
-
-### 🖋 Typography
-
-Typography tokens are available with **typography** property. Typography is provided with fontFamily, fontSize, lineHeight and fontWeight.
-
-Example from **tokens object**:
-
-```js
-exports.tokens = {
-  typography: {
-    heading1: {
-      fontFamily:
-        "'Source Sans Pro', 'Helvetica Neue', 'Arial', sans-serif",
-      fontSize: { value: 40, unit: 'px' },
-      lineHeight: { value: 48, unit: 'px' },
-      fontWeight: 300
+    whiteBase: 'hsl(0, 0%, 100%)'
+  },
+  values: {
+    colors: {
+      whiteBase: { h: 0, s: 0, l: 100 }
     }
   }
 };
@@ -156,33 +133,46 @@ exports.tokens = {
 JavaScript example:
 
 ```js
-const heading1 = tokens.typography.heading1;
-const heading1FontSize =
-  heading1.fontSize.value + heading1.fontSize.unit;
-const heading1FontFamily = heading1.fontFamily;
+const brandBaseLightness =
+  suomifiDesignTokens.values.colors.brandBase.l;
+
+const brandBaseCss = suomifiDesignTokens.colors.brandBase;
 ```
 
 TypeScript example with typings:
 
 ```ts
-import { TypographyToken } from 'suomifi-design-tokens';
+import { ColorToken } from 'suomifi-design-tokens';
 
-const heading1: TypographyToken = tokens.typography.heading1;
-const heading1FontSize: string =
-  heading1.fontSize.value + heading1.fontSize.unit;
-const heading1FontFamily: string = heading1.fontFamily;
+const brandBase: ColorToken =
+  suomifiDesignTokens.values.colors.brandBase;
+const brandBaseLightness: string = brandBase.l;
+
+const brandBaseCSS: string = suomifiDesignTokens.colors.brandBase;
 ```
 
-### 📏 Spacing
+### 🖋 Typography
 
-Spacing tokens are available with **spacing** property. Spacings are provided with numerical value and format string.
+Typography tokens are available with **typography** property as css strings. Separated fontFamily, fontSize, lineHeight and fontWeight properties can be accessed through **values** property.
 
-Example from **tokens object**:
+Excerpt from **suomifiDesignTokens** object:
 
 ```js
-exports.tokens = {
-  spacing: {
-    l: { value: 32, unit: 'px' }
+exports.suomifiDesignTokens = {
+  typography: {
+    heading1:
+      "font-family: 'Source Sans Pro', 'Helvetica Neue', 'Arial', sans-serif; font-size: 40px; line-height: 48px; font-weight: 300;"
+  },
+  values: {
+    typography: {
+      heading1: {
+        fontFamily:
+          "'Source Sans Pro', 'Helvetica Neue', 'Arial', sans-serif",
+        fontSize: { value: 40, unit: 'px' },
+        lineHeight: { value: 48, unit: 'px' },
+        fontWeight: 300
+      }
+    }
   }
 };
 ```
@@ -190,8 +180,54 @@ exports.tokens = {
 JavaScript example:
 
 ```js
-const spacingL = tokens.spacing.l;
+const heading1 = suomifiDesignTokens.values.typography.heading1;
+const heading1FontSize =
+  heading1.fontSize.value + heading1.fontSize.unit;
+const heading1FontFamily = heading1.fontFamily;
+
+const heading1Css = suomifiDesignTokens.typography.heading1;
+```
+
+TypeScript example with typings:
+
+```ts
+import { TypographyToken } from 'suomifi-design-tokens';
+
+const heading1: TypographyToken =
+  suomifiDesignTokens.values.typography.heading1;
+const heading1FontSize: string =
+  heading1.fontSize.value + heading1.fontSize.unit;
+const heading1FontFamily: string = heading1.fontFamily;
+
+const heading1Css: string = suomifiDesignTokens.typography.heading1;
+```
+
+### 📏 Spacing
+
+Spacing tokens are available with **spacing** property as CSS strings. Separated unit and value properties can be accessed through **values** property.
+
+Excerpt from **suomifiDesignTokens** object:
+
+```js
+exports.suomifiDesignTokens = {
+  spacing: {
+    l: '32px'
+  },
+  values: {
+    spacing: {
+      l: { value: 32, unit: 'px' }
+    }
+  }
+};
+```
+
+JavaScript example:
+
+```js
+const spacingL = suomifiDesignTokens.values.spacing.l;
 const spacingLWithUnit = spacingL.value + spacingL.unit;
+
+const spacingLCss = suomifiDesignTokens.spacing.l;
 ```
 
 TypeScript example with typings:
@@ -199,8 +235,10 @@ TypeScript example with typings:
 ```ts
 import { ValueUnit } from 'suomifi-design-tokens';
 
-const spacingL: ValueUnit = tokens.spacing.l;
+const spacingL: ValueUnit = suomifiDesignTokens.values.spacing.l;
 const spacingLWithUnit: string = spacingL.value + spacingL.unit;
+
+const spacingLCss: string = suomifiDesignTokens.spacing.l;
 ```
 
 ## ⌨️ Development
