--- src/converters/typescript-css-in-js/interfaces.ts.template
@@ -1,5 +1,0 @@
-export interface DesignTokens {
-  colors: ColorDesignTokens;
-  typography: TypographyDesignTokens;
-  spacing: SpacingDesignTokens;
-}
